Fixing bugs.

+++ libpimeval/src/pimCmd.h
@@ -152,7 +152,7 @@ protected:
     case PIM_UINT16: operandValue =  *reinterpret_cast<uint16_t*>(&operandBits); break;
     case PIM_UINT32: operandValue =  *reinterpret_cast<uint32_t*>(&operandBits); break;
     case PIM_UINT64: operandValue =  *reinterpret_cast<uint64_t*>(&operandBits); break;
-    case PIM_FP32: operandValue =  *reinterpret_cast<uint64_t*>(&operandBits); break;
+    case PIM_FP32: operandValue =  *reinterpret_cast<float*>(&operandBits); break;
     default:
         std::printf("PIM-Error: Unsupported data type %u\n", static_cast<unsigned>(dataType));
     }
